Remove old title sequence code
