+++ openrct2.vcxproj
@@ -151,7 +151,6 @@
     <ClCompile Include="src\interface\graph.c" />
     <ClCompile Include="src\interface\keyboard_shortcut.c" />
     <ClCompile Include="src\interface\screenshot.c" />
-    <ClCompile Include="src\interface\title_sequences.c" />
     <ClCompile Include="src\interface\viewport.c" />
     <ClCompile Include="src\interface\viewport_interaction.c" />
     <ClCompile Include="src\interface\widget.c" />
@@ -488,7 +487,6 @@
     <ClInclude Include="src\interface\graph.h" />
     <ClInclude Include="src\interface\keyboard_shortcut.h" />
     <ClInclude Include="src\interface\screenshot.h" />
-    <ClInclude Include="src\interface\title_sequences.h" />
     <ClInclude Include="src\interface\viewport.h" />
     <ClInclude Include="src\interface\widget.h" />
     <ClInclude Include="src\interface\window.h" />
