+++ src/windows/loadsave.c
@@ -19,12 +19,12 @@
 #include "../editor.h"
 #include "../game.h"
 #include "../interface/themes.h"
-#include "../interface/title_sequences.h"
 #include "../interface/widget.h"
 #include "../interface/window.h"
 #include "../localisation/localisation.h"
 #include "../network/network.h"
 #include "../scenario.h"
+#include "../title/TitleScreen.h"
 #include "../util/util.h"
 #include "../windows/error.h"
