+++ src/windows/title_command_editor.c
@@ -17,7 +17,6 @@
 #include "../game.h"
 #include "../input.h"
 #include "../interface/themes.h"
-#include "../interface/title_sequences.h"
 #include "../interface/viewport.h"
 #include "../interface/widget.h"
 #include "../interface/window.h"
